travis: force nodejs version 12.x
Use 'nvm' to install and activate nodejs version 12.x
when running Travis-CI builds.

Previously we used some default version, which was really
old.

As per travis-ci issue comments:
https://github.com/travis-ci/travis-ci/issues/3661

the 'node_js' verion is ignored for project's that are
not pure nodejs projects. In our case, where we are
python and nodejs project, the recommended way is to
use 'nvm' commands to manage build's nodejs version.