fix entering name on the observer dialog
Use 'onChange' prop instead of 'inputRef' to access entered name
on the observer dialog.

The 'inputRef' does not seems to exist, with react throwing
this error:

"<..> React does not recognize the `inputRef` prop on a DOM element. <..>"

Entering custom name did not work. The login id was always used as nick,
regardless if anything was entered in the text box.