force signout other sessions
When a user is logging in 'force signout' the user from previous session.

This way, the user is not prevented from loggin in to the MxCube in cases where
there is some old session kicking around somewhere.

Send the socketio message to UI informing that the session is over _after_ the
session have been invalidated. This to prevent race bugs.

Also, renamed that message to 'signedOut', as it no longer used exclusivly
for kicking out observer users.