Fix MXCuBE dependencies
The dependencies `mxcubecore` and `mxcube-video-streamer`
were pinned to a specific version in `pyproject.toml`.
The version constraints in `pyproject.toml` are now more loose,
and the lockfile has been updated with the latest versions of those.

The command used to do this was:

```none
poetry add 'mxcubecore>=1.54.0' 'mxcube-video-streamer>=1.0.0'
```

GitHub: fix https://github.com/mxcube/mxcubeweb/issues/1159