UI: treat disabled state as closed state in InOutSwitch
Change logic for handling DISABLED state in the InOutSwitch
component. Make the DISABLED state equivalent to CLOSED state.

This solves the issue with the 'Sample Changer' InOutSwitch widget.
Previously it was not possible to power on a disabled sample
changer, as the command for DISABLED state was set to 'PowerOff'.
Now when it is in DISABLED state, the command will be 'PowerOn'.

Note that for sample changers, the DISABLED state is used when
they are in 'powered off' mode.